fix: 모듈 순환참조 오류
